requirements.txt
pydantic==2.7.1
typing-extensions
python-dotenv
pytest
pyyaml
fastapi
uvicorn
openai